+++ packages/form-schema/src/parse.ts
@@ -1,4 +1,5 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
+import isPromise from 'is-promise';
 
 import type {
   GetFieldSchema,
@@ -38,12 +39,15 @@ Errors extends Record<string, any>,
     SerializedValues,
     Errors
     >,
-  ): Values => {
+  ): Values | Promise<Values> => {
   if (!values) {
     return null;
   }
 
   const res = {} as Values;
+
+  let hasPromise = false;
+  const preparsedValues: Array<Values | Promise<Values>> = [];
 
   names.forEach((name) => {
     const fieldSchema = getFieldSchema(name);
@@ -60,13 +64,34 @@ Errors extends Record<string, any>,
       )
       : getFieldSchema;
 
-    Object.assign(res, parser(
+    const parserResult = parser(
       values,
       name,
       fieldSchema,
       computedGetFieldSchema,
       getFieldType,
-    ));
+    );
+
+    if (isPromise(parserResult)) {
+      hasPromise = true;
+    }
+
+    preparsedValues.push(parserResult);
+  });
+
+  if (hasPromise) {
+    return Promise.all(preparsedValues)
+      .then((parsedValues) => {
+        parsedValues.forEach((parsedValue: Values) => {
+          Object.assign(res, parsedValue);
+        });
+
+        return res;
+      });
+  }
+
+  preparsedValues.forEach((parsedValue: Values) => {
+    Object.assign(res, parsedValue);
   });
 
   return res;